Move stray concern for nilifying blanks
* Was in lib, can now live in Concerns::NilifyBlanks
