+++ app/models/concerns/nilify_blanks.rb
@@ -1,0 +1,15 @@
+module Concerns
+  module NilifyBlanks
+    extend ActiveSupport::Concern
+
+    included do
+      before_save :nilify_blanks
+    end
+
+    def nilify_blanks
+      attributes.each do |column, value|
+        self[column] = nil if value.blank?
+      end
+    end
+  end
+end
