+++ app/models/mapping.rb
@@ -1,9 +1,8 @@
-require 'active_record/concerns/nilify_blanks'
 require 'digest/sha1'
 require 'transition/history'
 
 class Mapping < ActiveRecord::Base
-  include ActiveRecord::Concerns::NilifyBlanks
+  include Concerns::NilifyBlanks
 
   # ActiveRecord uses a column named 'type' for Single Table Inheritance, and
   # by default activates STI if a 'type' column is present. Setting the column
