--- lib/active_record/concerns/nilify_blanks.rb
@@ -1,17 +1,0 @@
-module ActiveRecord
-  module Concerns
-    module NilifyBlanks
-      extend ActiveSupport::Concern
-
-      included do
-        before_save :nilify_blanks
-      end
-
-      def nilify_blanks
-        attributes.each do |column, value|
-          self[column] = nil if value.blank?
-        end
-      end
-    end
-  end
-end
